--- src/main/java/seedu/zerotoone/ui/StatusBarFooter.java
@@ -1,27 +1,0 @@
-package seedu.zerotoone.ui;
-
-import java.nio.file.Path;
-import java.nio.file.Paths;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-import javafx.scene.layout.Region;
-import seedu.zerotoone.ui.util.UiPart;
-
-/**
- * A ui for the status bar that is displayed at the footer of the application.
- */
-public class StatusBarFooter extends UiPart<Region> {
-
-    private static final String FXML = "StatusBarFooter.fxml";
-
-    @FXML
-    private Label saveLocationStatus;
-
-
-    public StatusBarFooter(Path saveLocation) {
-        super(FXML);
-        saveLocationStatus.setText(Paths.get(".").resolve(saveLocation).toString());
-    }
-
-}
