+++ src/main/resources/view/log/StatisticsWindow.fxml
